Removed unnecessary cast to double

+++ Mathmagician/PrimeNumbers.cs
@@ -20,7 +20,7 @@ namespace Mathmagician
             if (a == 2 || a == 3) return true;
             if (!((a & 1) == 1)) return false;
             if (!((a + 1) % 6 == 0 || (a - 1) % 6 == 0)) return false;
-            int q = (int)Math.Sqrt((double)a) +1;
+            int q = (int)Math.Sqrt(a) +1;
             for (int v = 3; v < q; v += 2)
                 if (a % v == 0)
                     return false;
